# hub/tests/golden/openapi.json
{
  "components": {
    "schemas": {
      "AgentCreate": {
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "slug": {
            "maxLength": 50,
            "pattern": "^[a-z][a-z0-9-]{1,48}[a-z0-9]$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "display_name"
        ],
        "title": "AgentCreate",
        "type": "object"
      },
      "AgentKeyCreate": {
        "properties": {
          "label": {
            "default": "",
            "maxLength": 120,
            "title": "Label",
            "type": "string"
          }
        },
        "title": "AgentKeyCreate",
        "type": "object"
      },
      "AgentKeyCreated": {
        "description": "Creation response — the only time the key appears in plaintext.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "key"
        ],
        "title": "AgentKeyCreated",
        "type": "object"
      },
      "AgentKeyOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "id",
          "label",
          "created_at",
          "revoked_at"
        ],
        "title": "AgentKeyOut",
        "type": "object"
      },
      "AgentOut": {
        "properties": {
          "allowed_senders": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Senders"
          },
          "auto_timeout_secs": {
            "title": "Auto Timeout Secs",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "instructions": {
            "title": "Instructions",
            "type": "string"
          },
          "max_auto_per_hour": {
            "title": "Max Auto Per Hour",
            "type": "integer"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "slug",
          "user_id",
          "display_name",
          "created_at",
          "mode",
          "allowed_senders",
          "max_auto_per_hour",
          "auto_timeout_secs",
          "instructions"
        ],
        "title": "AgentOut",
        "type": "object"
      },
      "AgentSettingsUpdate": {
        "description": "Partial PATCH — only fields that are present get changed.",
        "properties": {
          "allowed_senders": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Senders"
          },
          "auto_timeout_secs": {
            "anyOf": [
              {
                "maximum": 600.0,
                "minimum": 10.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Timeout Secs"
          },
          "clear_allowed_senders": {
            "default": false,
            "title": "Clear Allowed Senders",
            "type": "boolean"
          },
          "instructions": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "max_auto_per_hour": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Auto Per Hour"
          },
          "mode": {
            "anyOf": [
              {
                "pattern": "^(inbox|auto)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          }
        },
        "title": "AgentSettingsUpdate",
        "type": "object"
      },
      "BroadcastRequest": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "from": {
            "maxLength": 50,
            "minLength": 3,
            "title": "From",
            "type": "string"
          },
          "group": {
            "maxLength": 51,
            "minLength": 2,
            "pattern": "^@[a-z][a-z0-9-]*$",
            "title": "Group",
            "type": "string"
          },
          "priority": {
            "default": "normal",
            "pattern": "^(urgent|high|normal|low)$",
            "title": "Priority",
            "type": "string"
          },
          "type": {
            "default": "request",
            "pattern": "^(request|response|notification|task|alert|status|ack)$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "from",
          "group",
          "body"
        ],
        "title": "BroadcastRequest",
        "type": "object"
      },
      "BroadcastResult": {
        "description": "Result of a fan-out: who received it, who was skipped (allowlist).",
        "properties": {
          "group": {
            "title": "Group",
            "type": "string"
          },
          "message_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Message Ids",
            "type": "array"
          },
          "sent": {
            "items": {
              "type": "string"
            },
            "title": "Sent",
            "type": "array"
          },
          "skipped": {
            "items": {
              "type": "string"
            },
            "title": "Skipped",
            "type": "array"
          }
        },
        "required": [
          "group",
          "sent",
          "skipped",
          "message_ids"
        ],
        "title": "BroadcastResult",
        "type": "object"
      },
...
      "DirectoryEntry": {
        "description": "Public view of an agent for the team directory.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "online": {
            "title": "Online",
            "type": "boolean"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "display_name",
          "online"
        ],
        "title": "DirectoryEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageOut": {
        "description": "A message as it travels over REST and WS (the `from` field is an alias).",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "from": {
            "title": "From",
            "type": "string"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "in_reply_to": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Reply To"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "to": {
            "title": "To",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "from",
          "to",
          "body",
          "type",
          "priority",
          "thread_id",
          "in_reply_to",
          "created_at",
          "delivered_at",
          "expires_at"
        ],
        "title": "MessageOut",
        "type": "object"
      },
...
      "SendMessageRequest": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "from": {
            "maxLength": 50,
            "minLength": 3,
            "title": "From",
            "type": "string"
          },
          "in_reply_to": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Reply To"
          },
          "priority": {
            "default": "normal",
            "pattern": "^(urgent|high|normal|low)$",
            "title": "Priority",
            "type": "string"
          },
          "to": {
            "maxLength": 50,
            "minLength": 3,
            "title": "To",
            "type": "string"
          },
          "type": {
            "default": "request",
            "pattern": "^(request|response|notification|task|alert|status|ack)$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "from",
          "to",
          "body"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Ampla Hub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agents": {
      "get": {
        "operationId": "list_my_agents_api_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentOut"
                  },
                  "title": "Response List My Agents Api Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Agents",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_agent_api_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/directory": {
      "get": {
        "operationId": "directory_api_agents_directory_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DirectoryEntry"
                  },
                  "title": "Response Directory Api Agents Directory Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Directory",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{slug}": {
      "get": {
        "operationId": "get_agent_api_agents__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{slug}/keys": {
      "get": {
        "operationId": "list_keys_api_agents__slug__keys_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentKeyOut"
                  },
                  "title": "Response List Keys Api Agents  Slug  Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Keys",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_key_api_agents__slug__keys_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentKeyCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Key",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{slug}/keys/{key_id}": {
      "delete": {
        "operationId": "revoke_key_api_agents__slug__keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentKeyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Key",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{slug}/settings": {
      "patch": {
        "operationId": "update_settings_api_agents__slug__settings_patch",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Settings",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/messages": {
      "post": {
        "description": "Panel: a human sends on behalf of their own agent. Real-time delivery\nmirrors the WS flow (delivered + mirror to observers).",
        "operationId": "send_message_api_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages/broadcast": {
      "post": {
        "description": "Panel: fan-out to @group or @all on behalf of the agent's own owner.",
        "operationId": "broadcast_api_messages_broadcast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Broadcast",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages/conversation": {
      "get": {
        "operationId": "conversation_api_messages_conversation_get",
        "parameters": [
          {
            "in": "query",
            "name": "a",
            "required": true,
            "schema": {
              "maxLength": 50,
              "minLength": 3,
              "title": "A",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "b",
            "required": true,
            "schema": {
              "maxLength": 50,
              "minLength": 3,
              "title": "B",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageOut"
                  },
                  "title": "Response Conversation Api Messages Conversation Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Conversation",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages/partners": {
      "get": {
        "operationId": "partners_api_messages_partners_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent",
            "required": true,
            "schema": {
              "maxLength": 50,
              "minLength": 3,
              "title": "Agent",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationPartner"
                  },
                  "title": "Response Partners Api Messages Partners Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Partners",
        "tags": [
          "messages"
        ]
      }
    },
...
  }
}